biginteger: fix overflow error from 64-bit internal data structure.
Initially `unsigned long` type is used as an internal data structure and
the type was considered as a 32-bit unsigned integer.

However, The size of signed/unsigned `long` type depends on the architecture
on which the program is running. If the program is on a 64-bit
architecture system, it will be a 64-bit integer which is the same with
`unsigned long long`.

Therefore, the internal `unsigned long` structure on 64-bit systems
cannot detect overflow errors.

The internal `unsigned long` type is changed to `uint32_t` which is the
same with `unsigned int` to prevent overflows and to express its size
explicitly.

+++ numeric/Integer.hpp
@@ -7,16 +7,17 @@
 #include <cctype>
 #include <vector>
 #include <string>
+#include <cstdint>
 
 class Integer {
 private:
 	bool _negative;
-	std::vector<unsigned long> _data;
+	std::vector<uint32_t> _data;
 
 private:
-	static const unsigned long long _base { std::numeric_limits<unsigned long>::max() };
+	static const uint64_t _base { std::numeric_limits<uint32_t>::max() };
 
-	static int _compare(const std::vector<unsigned long> &a, const std::vector<unsigned long> &b) {
+	static int _compare(const std::vector<uint32_t> &a, const std::vector<uint32_t> &b) {
 		if (a.size() != b.size())
 			return a.size() < b.size()? -1: 1;
 
@@ -34,7 +35,7 @@ public:
 
 	template <typename T, typename=typename std::enable_if<std::is_arithmetic<T>::value>::type>
 	Integer(const T &i): _negative(false) {
-		long long v = i;
+		int64_t v = i;
 		if (v < 0LL)
 			_negative = true, v = -v;
 
@@ -58,7 +59,7 @@ public:
 				if (!isdigit(*it))
 					throw std::runtime_error("Integer cannot be initialized from a non-number string: " + s);
 
-				unsigned long long val = *it - '0';
+				uint64_t val = *it - '0';
 				for (auto &d : _data) {
 					val += 10ULL * d;
 					d = val % _base;
